% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello
A function to say hello}
\usage{
hello(name)
}
\arguments{
\item{name}{Name of individual(s) to greet}
}
\value{
A character vector of greetings
}
\description{
hello
A function to say hello
}
\details{
This is a toy function that will return a greeting.
For example, \code{hello('BIFX 545 Students')} will return "Hello BIFX 545 Students!"
}
